+++ sdk/metricsadvisor/azure-ai-metricsadvisor/tests/base_testcase_aad.py
@@ -159,7 +159,7 @@ class TestMetricsAdvisorAdministrationClientBase(AzureTestCase):
             name=name,
             source=SqlServerDataFeedSource(
                 connection_string=self.sql_server_connection_string,
-                query="select * from adsample2 where Timestamp = @StartTime"
+                query="select * from adsample2 where Timestamp = @StartTime"
             ),
             granularity="Daily",
             schema=DataFeedSchema(
@@ -205,7 +205,7 @@ class TestMetricsAdvisorAdministrationClientBase(AzureTestCase):
             name=data_feed_name,
             source=SqlServerDataFeedSource(
                 connection_string=self.sql_server_connection_string,
-                query=u"select * from adsample2 where Timestamp = @StartTime"
+                query=u"select * from adsample2 where Timestamp = @StartTime"
             ),
             granularity=DataFeedGranularity(
                 granularity_type="Daily",