+++ sdk/monitor/azure-monitor-query/CHANGELOG.md
@@ -1,7 +1,23 @@
 # Release History
 
+## 1.0.0b4 (Unreleased)
 
-## 1.0.0b3 (Unreleased)
+### Features Added
+
+### Breaking Changes
+
+- Rename `batch_query` to `query_batch`.
+- Rename `LogsBatchQueryRequest` to `LogsBatchQuery`.
+- `include_render` is now renamed to `include_visualization` in the query API.
+- `LogsQueryResult` and `LogsBatchQueryResult` now return `visualization` instead of `render`.
+
+### Bugs Fixed
+
+- `include_statistics` and `include_visualization` args can now work together.
+
+### Other Changes
+
+## 1.0.0b3 (2021-08-09)
 
 ### Features Added
 
@@ -17,10 +33,6 @@
 - `LogsQueryRequest` is renamed to `LogsBatchQueryRequest`
 - `LogsQueryResults` is now renamed to `LogsQueryResult`
 - `LogsBatchQueryResult` now has 4 additional attributes - `tables`, `error`, `statistics` and `render` instead of `body` attribute.
-
-### Bugs Fixed
-
-### Other Changes
 
 ## 1.0.0b2 (2021-07-06)