+++ sdk/synapse/azure-synapse-managedprivateendpoints/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Updated API version to "2020-12-01" which is the default API version
 - Added API version "2021-06-01-preview" support
+- `ManagedPrivateEndpointsClient` is renamed to `VnetClient`
 
 ## 0.3.0 (2021-03-09)